Add files via upload
Apparently, for whatever reason, the Camera was seeing the images. Just that they were very, very, terribly small. 
Scale Camera scales the gameObjects up to match the size of the Screen. Can be a bit iffy though if full screen is not used.
New Board to match, so the old Board will be deleted.