Small change
PlayerStats now includes damageable boolean that tracks if the player can take damage or not. This is set to false when the player activates their shield. Stan will hook this up with EnemyBehavior so that enemy damage is negated when the player cant be damaged. Stats in PlayerStats that you all will be accessing for various purposes are now static variables.